Rework import-vars
Our internal version of import-vars now has modification support split out into
import-vars-with-mods. This makes import-vars signature compatible with the
potemkin version and allow it to be included in clj-kondo checks. Coolio!

This allowed me to find a couple of unintended deprecated usages of functions in
tests and made gives me a more complete clj-kondo config.