Provide access insert-child in underlying zipper from Zip API
The underlying zipper is used to operate on the zipper without skipping
whitespace. We provide access from the Zip API to the underlying zipper
functions via their * equivalents. For example zip/next* is the no whitespace
skipping equivalent of the whitespace skipping zip/next.

The missing zip/insert-child* is now also included.