Correct node? function and make public
I had previously factored out verification that an element
was a rewrite-cljc created node from tests to a "node?" function.

Thing is, the test was never correct.

Rewrite-clj, and therefore rewrite-cljc, extends Object to a Node.
I'm not entirely clear why it does this, but assume it is important.
This means any object will return true for `satisfies?` Node, which
is not helpful for a node? fn.

I added some tests for "node?" and updated it to return true only
for rewrite-clj created nodes.

I also decided to expose this function under the node API.